+++ engine/api/typescript/src/index.ts
@@ -2,6 +2,6 @@
 export { KURTOSIS_ENGINE_VERSION } from "./kurtosis_engine_version/kurtosis_engine_version";
 
 // TODO Remove this - shouldn't be necessary to be exported due to the newKurtosisContextFromLocalEngine() method
-export { KurtosisContext, DEFAULT_GRPC_PROXY_ENGINE_SERVER_PORT_NUM, DEFAULT_GRPC_ENGINE_SERVER_PORT_NUM } from "./lib/kurtosis_context/kurtosis_context";
+export { KurtosisContext, DEFAULT_GRPC_PROXY_ENGINE_SERVER_PORT_NUM, DEFAULT_GRPC_ENGINE_SERVER_PORT_NUM, DEFAULT_HTTP_LOGS_COLLECTOR_PORT_NUM } from "./lib/kurtosis_context/kurtosis_context";
 
 export { EnclaveAPIContainerHostMachineInfo } from "./kurtosis_engine_rpc_api_bindings/engine_service_pb"