+++ engine/api/typescript/src/lib/kurtosis_context/kurtosis_context.ts
@@ -29,8 +29,9 @@ const LOCAL_HOSTNAME: string = "localhost";
 
 const API_CONTAINER_LOG_LEVEL: string = "debug";
 
-export const DEFAULT_GRPC_PROXY_ENGINE_SERVER_PORT_NUM: number = 9711;
 export const DEFAULT_GRPC_ENGINE_SERVER_PORT_NUM: number = 9710;
+export const DEFAULT_GRPC_PROXY_ENGINE_SERVER_PORT_NUM: number = 9711;
+export const DEFAULT_HTTP_LOGS_COLLECTOR_PORT_NUM: number = 9712;
 
 // Blank tells the engine server to use the default
 const DEFAULT_API_CONTAINER_VERSION_TAG = "";