+++ engine/docs/changelog.md
@@ -1,5 +1,11 @@
 # TBD
+### Features
+* Published the `DefaultHttpLogsCollectorPortNum` const which can be used to launch the engine's server
 
+### Breaking Changes
+* Renamed the `DefaultKurtosisEngineServerGrpcPortNum` const to `DefaultGrpcEngineServerPortNum` and `DefaultKurtosisEngineServerGrpcProxyPortNum` to `DefaultGrpcProxyEngineServerPortNum` in Golang library under our principle of TS & Go being as similar as possible
+  * Users will have to replace the old const with the new ones, the values does not change
+ 
 # 1.33.0
 ### Changes
 * Upgraded core to 1.59.1 and container-engine-lib to 0.38.0